Struct google_genomics1::Annotation

pub struct Annotation {
    pub info: Option<HashMap<String, Vec<String>>>,
    pub reference_id: Option<String>,
    pub end: Option<String>,
    pub name: Option<String>,
    pub type_: Option<String>,
    pub variant: Option<VariantAnnotation>,
    pub start: Option<String>,
    pub annotation_set_id: Option<String>,
    pub reference_name: Option<String>,
    pub reverse_strand: Option<bool>,
    pub transcript: Option<Transcript>,
    pub id: Option<String>,
}

An annotation describes a region of reference genome. The value of an annotation may be one of several canonical types, supplemented by arbitrary info tags. An annotation is not inherently associated with a specific sample or individual (though a client could choose to use annotations in this way). Example canonical annotation types are GENE and VARIANT.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

search annotations (none)

create annotations (request|response)

delete annotations (none)

update annotations (request|response)

get annotations (response)

batch create annotations (none)

Fields

info

A map of additional read alignment information. This must be of the form map (string key mapping to a list of string values).

reference_id

The ID of the Google Genomics reference associated with this range.

end

The end position of the range on the reference, 0-based exclusive.

name

The display name of this annotation.

type_

The data type for this annotation. Must match the containing annotation set's type.

variant

A variant annotation, which describes the effect of a variant on the genome, the coding sequence, and/or higher level consequences at the organism level e.g. pathogenicity. This field is only set for annotations of type VARIANT.

start

The start position of the range on the reference, 0-based inclusive.

annotation_set_id

The annotation set to which this annotation belongs.

reference_name

The display name corresponding to the reference specified by referenceId, for example chr1, 1, or chrX.

reverse_strand

Whether this range refers to the reverse strand, as opposed to the forward strand. Note that regardless of this field, the start/end position of the range always refer to the forward strand.

transcript

A transcript value represents the assertion that a particular region of the reference genome may be transcribed as RNA. An alternative splicing pattern would be represented as a separate transcript object. This field is only set for annotations of type TRANSCRIPT.

id

The server-generated annotation ID, unique across all annotations.