Struct google_genomics1::AnnotationsetMethods

pub struct AnnotationsetMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on annotationset resources. It is not used directly, but through the Genomics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_genomics1 as genomics1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use genomics1::Genomics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Genomics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create(...)`, `delete(...)`, `get(...)`, `search(...)` and `update(...)`
// to build up your call.
let rb = hub.annotationsets();

Methods

impl<'a, C, A> AnnotationsetMethods<'a, C, A>

fn update(&self, request: AnnotationSet, annotation_set_id: &str) -> AnnotationsetUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an annotation set. The update must respect all mutability restrictions and other invariants described on the annotation set resource. Caller must have WRITE permission for the associated dataset.

Arguments

request - No description provided.

annotationSetId - The ID of the annotation set to be updated.

fn search(&self, request: SearchAnnotationSetsRequest) -> AnnotationsetSearchCall<'a, C, A>

Create a builder to help you perform the following task:

Searches for annotation sets that match the given criteria. Annotation sets are returned in an unspecified order. This order is consistent, such that two queries for the same content (regardless of page size) yield annotation sets in the same order across their respective streams of paginated responses. Caller must have READ permission for the queried datasets.

Arguments

request - No description provided.

fn create(&self, request: AnnotationSet) -> AnnotationsetCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new annotation set. Caller must have WRITE permission for the associated dataset. The following fields are required: * datasetId * referenceSetId All other fields may be optionally specified, unless documented as being server-generated (for example, the id field).

Arguments

request - No description provided.

fn delete(&self, annotation_set_id: &str) -> AnnotationsetDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an annotation set. Caller must have WRITE permission for the associated annotation set.

Arguments

annotationSetId - The ID of the annotation set to be deleted.

fn get(&self, annotation_set_id: &str) -> AnnotationsetGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets an annotation set. Caller must have READ permission for the associated dataset.

Arguments

annotationSetId - The ID of the annotation set to be retrieved.