Struct google_genomics1::CallsetMethods

pub struct CallsetMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on callset resources. It is not used directly, but through the Genomics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_genomics1 as genomics1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use genomics1::Genomics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Genomics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create(...)`, `delete(...)`, `get(...)`, `patch(...)` and `search(...)`
// to build up your call.
let rb = hub.callsets();

Methods

impl<'a, C, A> CallsetMethods<'a, C, A>

fn search(&self, request: SearchCallSetsRequest) -> CallsetSearchCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a list of call sets matching the criteria. For the definitions of call sets and other genomics resources, see Fundamentals of Google Genomics Implements GlobalAllianceApi.searchCallSets.

Arguments

request - No description provided.

fn patch(&self, request: CallSet, call_set_id: &str) -> CallsetPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a call set. For the definitions of call sets and other genomics resources, see Fundamentals of Google Genomics This method supports patch semantics.

Arguments

request - No description provided.

callSetId - The ID of the call set to be updated.

fn create(&self, request: CallSet) -> CallsetCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new call set. For the definitions of call sets and other genomics resources, see Fundamentals of Google Genomics

Arguments

request - No description provided.

fn delete(&self, call_set_id: &str) -> CallsetDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a call set. For the definitions of call sets and other genomics resources, see Fundamentals of Google Genomics

Arguments

callSetId - The ID of the call set to be deleted.

fn get(&self, call_set_id: &str) -> CallsetGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a call set by ID. For the definitions of call sets and other genomics resources, see Fundamentals of Google Genomics

Arguments

callSetId - The ID of the call set.