Struct google_genomics1::DatasetMethods

pub struct DatasetMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on dataset resources. It is not used directly, but through the Genomics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_genomics1 as genomics1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use genomics1::Genomics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Genomics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create(...)`, `delete(...)`, `get(...)`, `get_iam_policy(...)`, `list(...)`, `patch(...)`, `set_iam_policy(...)`, `test_iam_permissions(...)` and `undelete(...)`
// to build up your call.
let rb = hub.datasets();

Methods

impl<'a, C, A> DatasetMethods<'a, C, A>

fn create(&self, request: Dataset) -> DatasetCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new dataset. For the definitions of datasets and other genomics resources, see Fundamentals of Google Genomics

Arguments

request - No description provided.

fn delete(&self, dataset_id: &str) -> DatasetDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a dataset and all of its contents (all read group sets, reference sets, variant sets, call sets, annotation sets, etc.) This is reversible (up to one week after the deletion) via the datasets.undelete operation. For the definitions of datasets and other genomics resources, see Fundamentals of Google Genomics

Arguments

datasetId - The ID of the dataset to be deleted.

fn undelete(&self, request: UndeleteDatasetRequest, dataset_id: &str) -> DatasetUndeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Undeletes a dataset by restoring a dataset which was deleted via this API. For the definitions of datasets and other genomics resources, see Fundamentals of Google Genomics This operation is only possible for a week after the deletion occurred.

Arguments

request - No description provided.

datasetId - The ID of the dataset to be undeleted.

fn set_iam_policy(&self, request: SetIamPolicyRequest, resource: &str) -> DatasetSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy on the specified dataset. Replaces any existing policy. For the definitions of datasets and other genomics resources, see Fundamentals of Google Genomics See Setting a Policy for more information.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which policy is being specified. Format is datasets/.

fn test_iam_permissions(&self, request: TestIamPermissionsRequest, resource: &str) -> DatasetTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. See Testing Permissions for more information. For the definitions of datasets and other genomics resources, see Fundamentals of Google Genomics

Arguments

request - No description provided.

resource - REQUIRED: The resource for which policy is being specified. Format is datasets/.

fn get(&self, dataset_id: &str) -> DatasetGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a dataset by ID. For the definitions of datasets and other genomics resources, see Fundamentals of Google Genomics

Arguments

datasetId - The ID of the dataset.

fn list(&self) -> DatasetListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists datasets within a project. For the definitions of datasets and other genomics resources, see Fundamentals of Google Genomics

fn get_iam_policy(&self, request: GetIamPolicyRequest, resource: &str) -> DatasetGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy for the dataset. This is empty if the policy or resource does not exist. See Getting a Policy for more information. For the definitions of datasets and other genomics resources, see Fundamentals of Google Genomics

Arguments

request - No description provided.

resource - REQUIRED: The resource for which policy is being specified. Format is datasets/.

fn patch(&self, request: Dataset, dataset_id: &str) -> DatasetPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a dataset. For the definitions of datasets and other genomics resources, see Fundamentals of Google Genomics This method supports patch semantics.

Arguments

request - No description provided.

datasetId - The ID of the dataset to be updated.