Struct google_genomics1::Experiment

pub struct Experiment {
    pub sequencing_center: Option<String>,
    pub library_id: Option<String>,
    pub instrument_model: Option<String>,
    pub platform_unit: Option<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

sequencing_center

The sequencing center used as part of this experiment.

library_id

A client-supplied library identifier; a library is a collection of DNA fragments which have been prepared for sequencing from a sample. This field is important for quality control as error or bias can be introduced during sample preparation.

instrument_model

The instrument model used as part of this experiment. This maps to sequencing technology in the SAM spec.

platform_unit

The platform unit used as part of this experiment, for example flowcell-barcode.lane for Illumina or slide for SOLiD. Corresponds to the @RG PU field in the SAM spec.