Struct google_genomics1::ExportVariantSetRequest

pub struct ExportVariantSetRequest {
    pub project_id: Option<String>,
    pub format: Option<String>,
    pub bigquery_dataset: Option<String>,
    pub call_set_ids: Option<Vec<String>>,
    pub bigquery_table: Option<String>,
}

The variant data export request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

export variantsets (request)

Fields

project_id

Required. The Google Cloud project ID that owns the destination BigQuery dataset. The caller must have WRITE access to this project. This project will also own the resulting export job.

format

The format for the exported data.

bigquery_dataset

Required. The BigQuery dataset to export data to. This dataset must already exist. Note that this is distinct from the Genomics concept of "dataset".

call_set_ids

If provided, only variant call information from the specified call sets will be exported. By default all variant calls are exported.

bigquery_table

Required. The BigQuery table to export data to. If the table doesn't exist, it will be created. If it already exists, it will be overwritten.