Struct google_genomics1::LinearAlignment

pub struct LinearAlignment {
    pub position: Option<Position>,
    pub cigar: Option<Vec<CigarUnit>>,
    pub mapping_quality: Option<i32>,
}

A linear alignment can be represented by one CIGAR string. Describes the mapped position and local alignment of the read to the reference.

This type is not used in any activity, and only used as part of another schema.

Fields
position: Option<Position>

The position of this alignment.

cigar: Option<Vec<CigarUnit>>

Represents the local alignment of this sequence (alignment matches, indels, etc) against the reference.

mapping_quality: Option<i32>

The mapping quality of this alignment. Represents how likely the read maps to this position as opposed to other locations. Specifically, this is -10 log10 Pr(mapping position is wrong), rounded to the nearest integer.