Struct google_genomics1::Program

pub struct Program {
    pub command_line: Option<String>,
    pub version: Option<String>,
    pub id: Option<String>,
    pub name: Option<String>,
    pub prev_program_id: Option<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

command_line

The command line used to run this program.

version

The version of the program run.

id

The user specified locally unique ID of the program. Used along with prevProgramId to define an ordering between programs.

name

The display name of the program. This is typically the colloquial name of the tool used, for example 'bwa' or 'picard'.

prev_program_id

The ID of the program run before this one.