Struct google_genomics1::ReadGroup

pub struct ReadGroup {
    pub info: Option<HashMap<String, Vec<String>>>,
    pub predicted_insert_size: Option<i32>,
    pub description: Option<String>,
    pub programs: Option<Vec<Program>>,
    pub sample_id: Option<String>,
    pub experiment: Option<Experiment>,
    pub reference_set_id: Option<String>,
    pub id: Option<String>,
    pub dataset_id: Option<String>,
    pub name: Option<String>,
}

A read group is all the data that's processed the same way by the sequencer.

This type is not used in any activity, and only used as part of another schema.

Fields
info: Option<HashMap<String, Vec<String>>>

A map of additional read group information. This must be of the form map (string key mapping to a list of string values).

predicted_insert_size: Option<i32>

The predicted insert size of this read group. The insert size is the length the sequenced DNA fragment from end-to-end, not including the adapters.

description: Option<String>

A free-form text description of this read group.

programs: Option<Vec<Program>>

The programs used to generate this read group. Programs are always identical for all read groups within a read group set. For this reason, only the first read group in a returned set will have this field populated.

sample_id: Option<String>

A client-supplied sample identifier for the reads in this read group.

experiment: Option<Experiment>

The experiment used to generate this read group.

reference_set_id: Option<String>

The reference set the reads in this read group are aligned to.

id: Option<String>

The server-generated read group ID, unique for all read groups. Note: This is different than the @RG ID field in the SAM spec. For that value, see name.

dataset_id: Option<String>

The dataset to which this read group belongs.

name: Option<String>

The read group name. This corresponds to the @RG ID field in the SAM spec.