Struct google_genomics1::ReadgroupsetMethods

pub struct ReadgroupsetMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on readgroupset resources. It is not used directly, but through the Genomics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_genomics1 as genomics1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use genomics1::Genomics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Genomics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `coveragebuckets_list(...)`, `delete(...)`, `export(...)`, `get(...)`, `import(...)`, `patch(...)` and `search(...)`
// to build up your call.
let rb = hub.readgroupsets();

Methods

impl<'a, C, A> ReadgroupsetMethods<'a, C, A>

fn export(&self, request: ExportReadGroupSetRequest, read_group_set_id: &str) -> ReadgroupsetExportCall<'a, C, A>

Create a builder to help you perform the following task:

Exports a read group set to a BAM file in Google Cloud Storage. For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics Note that currently there may be some differences between exported BAM files and the original BAM file at the time of import. See ImportReadGroupSets for caveats.

Arguments

request - No description provided.

readGroupSetId - Required. The ID of the read group set to export. The caller must have READ access to this read group set.

fn import(&self, request: ImportReadGroupSetsRequest) -> ReadgroupsetImportCall<'a, C, A>

Create a builder to help you perform the following task:

Creates read group sets by asynchronously importing the provided information. For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics The caller must have WRITE permissions to the dataset. ## Notes on BAM import - Tags will be converted to strings - tag types are not preserved - Comments (@CO) in the input file header will not be preserved - Original header order of references (@SQ) will not be preserved - Any reverse stranded unmapped reads will be reverse complemented, and their qualities (also the "BQ" and "OQ" tags, if any) will be reversed - Unmapped reads will be stripped of positional information (reference name and position)

Arguments

request - No description provided.

fn patch(&self, request: ReadGroupSet, read_group_set_id: &str) -> ReadgroupsetPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a read group set. For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics This method supports patch semantics.

Arguments

request - No description provided.

readGroupSetId - The ID of the read group set to be updated. The caller must have WRITE permissions to the dataset associated with this read group set.

fn get(&self, read_group_set_id: &str) -> ReadgroupsetGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a read group set by ID. For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics

Arguments

readGroupSetId - The ID of the read group set.

fn coveragebuckets_list(&self, read_group_set_id: &str) -> ReadgroupsetCoveragebucketListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists fixed width coverage buckets for a read group set, each of which correspond to a range of a reference sequence. Each bucket summarizes coverage information across its corresponding genomic range. For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics Coverage is defined as the number of reads which are aligned to a given base in the reference sequence. Coverage buckets are available at several precomputed bucket widths, enabling retrieval of various coverage 'zoom levels'. The caller must have READ permissions for the target read group set.

Arguments

readGroupSetId - Required. The ID of the read group set over which coverage is requested.

fn delete(&self, read_group_set_id: &str) -> ReadgroupsetDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a read group set. For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics

Arguments

readGroupSetId - The ID of the read group set to be deleted. The caller must have WRITE permissions to the dataset associated with this read group set.

fn search(&self, request: SearchReadGroupSetsRequest) -> ReadgroupsetSearchCall<'a, C, A>

Create a builder to help you perform the following task:

Searches for read group sets matching the criteria. For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics Implements GlobalAllianceApi.searchReadGroupSets.

Arguments

request - No description provided.