Struct google_genomics1::ReferenceMethods

pub struct ReferenceMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on reference resources. It is not used directly, but through the Genomics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_genomics1 as genomics1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use genomics1::Genomics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Genomics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `bases_list(...)`, `get(...)` and `search(...)`
// to build up your call.
let rb = hub.references();

Methods

impl<'a, C, A> ReferenceMethods<'a, C, A>

fn bases_list(&self, reference_id: &str) -> ReferenceBaseListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the bases in a reference, optionally restricted to a range. For the definitions of references and other genomics resources, see Fundamentals of Google Genomics Implements GlobalAllianceApi.getReferenceBases.

Arguments

referenceId - The ID of the reference.

fn search(&self, request: SearchReferencesRequest) -> ReferenceSearchCall<'a, C, A>

Create a builder to help you perform the following task:

Searches for references which match the given criteria. For the definitions of references and other genomics resources, see Fundamentals of Google Genomics Implements GlobalAllianceApi.searchReferences.

Arguments

request - No description provided.

fn get(&self, reference_id: &str) -> ReferenceGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a reference. For the definitions of references and other genomics resources, see Fundamentals of Google Genomics Implements GlobalAllianceApi.getReference.

Arguments

referenceId - The ID of the reference.