Struct google_genomics1::SearchReadGroupSetsRequest

pub struct SearchReadGroupSetsRequest {
    pub page_token: Option<String>,
    pub dataset_ids: Option<Vec<String>>,
    pub name: Option<String>,
    pub page_size: Option<i32>,
}

The read group set search request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

search readgroupsets (request)

Fields

page_token

The continuation token, which is used to page through large result sets. To get the next page of results, set this parameter to the value of nextPageToken from the previous response.

dataset_ids

Restricts this query to read group sets within the given datasets. At least one ID must be provided.

name

Only return read group sets for which a substring of the name matches this string.

page_size

The maximum number of results to return in a single page. If unspecified, defaults to 256. The maximum value is 1024.