Struct google_genomics1::SearchReadsRequest

pub struct SearchReadsRequest {
    pub start: Option<String>,
    pub read_group_set_ids: Option<Vec<String>>,
    pub read_group_ids: Option<Vec<String>>,
    pub page_size: Option<i32>,
    pub page_token: Option<String>,
    pub end: Option<String>,
    pub reference_name: Option<String>,
}

The read search request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

search reads (request)

Fields

start

The start position of the range on the reference, 0-based inclusive. If specified, referenceName must also be specified.

read_group_set_ids

The IDs of the read groups sets within which to search for reads. All specified read group sets must be aligned against a common set of reference sequences; this defines the genomic coordinates for the query. Must specify one of readGroupSetIds or readGroupIds.

read_group_ids

The IDs of the read groups within which to search for reads. All specified read groups must belong to the same read group sets. Must specify one of readGroupSetIds or readGroupIds.

page_size

The maximum number of results to return in a single page. If unspecified, defaults to 256. The maximum value is 2048.

page_token

The continuation token, which is used to page through large result sets. To get the next page of results, set this parameter to the value of nextPageToken from the previous response.

end

The end position of the range on the reference, 0-based exclusive. If specified, referenceName must also be specified.

reference_name

The reference sequence name, for example chr1, 1, or chrX. If set to *, only unmapped reads are returned. If unspecified, all reads (mapped and unmapped) are returned.