Struct google_genomics1::SearchVariantsRequest

pub struct SearchVariantsRequest {
    pub end: Option<String>,
    pub page_size: Option<i32>,
    pub page_token: Option<String>,
    pub max_calls: Option<i32>,
    pub start: Option<String>,
    pub call_set_ids: Option<Vec<String>>,
    pub variant_name: Option<String>,
    pub reference_name: Option<String>,
    pub variant_set_ids: Option<Vec<String>>,
}

The variant search request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

search variants (request)

Fields

end

The end of the window, 0-based exclusive. If unspecified or 0, defaults to the length of the reference.

page_size

The maximum number of variants to return in a single page. If unspecified, defaults to 5000. The maximum value is 10000.

page_token

The continuation token, which is used to page through large result sets. To get the next page of results, set this parameter to the value of nextPageToken from the previous response.

max_calls

The maximum number of calls to return in a single page. Note that this limit may be exceeded; at least one variant is always returned per page, even if it has more calls than this limit. If unspecified, defaults to 5000. The maximum value is 10000.

start

The beginning of the window (0-based, inclusive) for which overlapping variants should be returned. If unspecified, defaults to 0.

call_set_ids

Only return variant calls which belong to call sets with these ids. Leaving this blank returns all variant calls. If a variant has no calls belonging to any of these call sets, it won't be returned at all. Currently, variants with no calls from any call set will never be returned.

variant_name

Only return variants which have exactly this name.

reference_name

Required. Only return variants in this reference sequence.

variant_set_ids

At most one variant set ID must be provided. Only variants from this variant set will be returned. If omitted, a call set id must be included in the request.