Struct google_genomics1::StreamReadsRequest

pub struct StreamReadsRequest {
    pub start: Option<String>,
    pub end: Option<String>,
    pub total_shards: Option<i32>,
    pub read_group_set_id: Option<String>,
    pub reference_name: Option<String>,
    pub project_id: Option<String>,
    pub shard: Option<i32>,
}

The stream reads request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

stream reads (request)

Fields
start: Option<String>

The start position of the range on the reference, 0-based inclusive. If specified, referenceName must also be specified.

end: Option<String>

The end position of the range on the reference, 0-based exclusive. If specified, referenceName must also be specified.

total_shards: Option<i32>

Specifying totalShards causes a disjoint subset of the normal response payload to be returned for each query with a unique shard parameter specified. A best effort is made to yield equally sized shards. Sharding can be used to distribute processing amongst workers, where each worker is assigned a unique shard number and all workers specify the same totalShards number. The union of reads returned for all sharded queries [0, totalShards) is equal to those returned by a single unsharded query. Queries for different values of totalShards with common divisors will share shard boundaries. For example, streaming shard 2 of 5 totalShards yields the same results as streaming shards 4 and 5 of 10 totalShards. This property can be leveraged for adaptive retries.

read_group_set_id: Option<String>

The ID of the read group set from which to stream reads.

reference_name: Option<String>

The reference sequence name, for example chr1, 1, or chrX. If set to *, only unmapped reads are returned.

project_id: Option<String>

The Google Cloud project ID which will be billed for this access. The caller must have WRITE access to this project. Required.

shard: Option<i32>

Restricts results to a shard containing approximately 1/totalShards of the normal response payload for this query. Results from a sharded request are disjoint from those returned by all queries which differ only in their shard parameter. A shard may yield 0 results; this is especially likely for large values of totalShards. Valid values are [0, totalShards).