Struct google_genomics1::Transcript

pub struct Transcript {
    pub coding_sequence: Option<CodingSequence>,
    pub gene_id: Option<String>,
    pub exons: Option<Vec<Exon>>,
}

A transcript represents the assertion that a particular region of the reference genome may be transcribed as RNA.

This type is not used in any activity, and only used as part of another schema.

Fields

coding_sequence

The range of the coding sequence for this transcript, if any. To determine the exact ranges of coding sequence, intersect this range with those of the exons, if any. If there are any exons, the codingSequence must start and end within them. Note that in some cases, the reference genome will not exactly match the observed mRNA transcript e.g. due to variance in the source genome from reference. In these cases, exon.frame will not necessarily match the expected reference reading frame and coding exon reference bases cannot necessarily be concatenated to produce the original transcript mRNA.

gene_id

The annotation ID of the gene from which this transcript is transcribed.

exons

The exons that compose this transcript. This field should be unset for genomes where transcript splicing does not occur, for example prokaryotes. Introns are regions of the transcript that are not included in the spliced RNA product. Though not explicitly modeled here, intron ranges can be deduced; all regions of this transcript that are not exons are introns. Exonic sequences do not necessarily code for a translational product (amino acids). Only the regions of exons bounded by the codingSequence correspond to coding DNA sequence. Exons are ordered by start position and may not overlap.