Struct google_genomics1::Variant

pub struct Variant {
    pub info: Option<HashMap<String, Vec<String>>>,
    pub variant_set_id: Option<String>,
    pub end: Option<String>,
    pub calls: Option<Vec<VariantCall>>,
    pub created: Option<String>,
    pub id: Option<String>,
    pub filter: Option<Vec<String>>,
    pub start: Option<String>,
    pub names: Option<Vec<String>>,
    pub alternate_bases: Option<Vec<String>>,
    pub reference_name: Option<String>,
    pub quality: Option<f64>,
    pub reference_bases: Option<String>,
}

A variant represents a change in DNA sequence relative to a reference sequence. For example, a variant could represent a SNP or an insertion. Variants belong to a variant set. For more genomics resource definitions, see Fundamentals of Google Genomics Each of the calls on a variant represent a determination of genotype with respect to that variant. For example, a call might assign probability of 0.32 to the occurrence of a SNP named rs1234 in a sample named NA12345. A call belongs to a call set, which contains related calls typically from one sample.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch variants (request|response)

delete variants (none)

create variants (request|response)

get variants (response)

import variants (none)

merge variants (none)

search variants (none)

stream variants (none)

Fields
info: Option<HashMap<String, Vec<String>>>

A map of additional variant information. This must be of the form map (string key mapping to a list of string values).

variant_set_id: Option<String>

The ID of the variant set this variant belongs to.

end: Option<String>

The end position (0-based) of this variant. This corresponds to the first base after the last base in the reference allele. So, the length of the reference allele is (end - start). This is useful for variants that don't explicitly give alternate bases, for example large deletions.

calls: Option<Vec<VariantCall>>

The variant calls for this particular variant. Each one represents the determination of genotype with respect to this variant.

created: Option<String>

The date this variant was created, in milliseconds from the epoch.

id: Option<String>

The server-generated variant ID, unique across all variants.

filter: Option<Vec<String>>

A list of filters (normally quality filters) this variant has failed. PASS indicates this variant has passed all filters.

start: Option<String>

The position at which this variant occurs (0-based). This corresponds to the first base of the string of reference bases.

names: Option<Vec<String>>

Names for the variant, for example a RefSNP ID.

alternate_bases: Option<Vec<String>>

The bases that appear instead of the reference bases.

reference_name: Option<String>

The reference on which this variant occurs. (such as chr20 or X)

quality: Option<f64>

A measure of how likely this variant is to be real. A higher value is better.

reference_bases: Option<String>

The reference bases for this variant. They start at the given position.