Struct google_genomics1::VariantCall

pub struct VariantCall {
    pub info: Option<HashMap<String, Vec<String>>>,
    pub genotype: Option<Vec<i32>>,
    pub call_set_id: Option<String>,
    pub phaseset: Option<String>,
    pub genotype_likelihood: Option<Vec<f64>>,
    pub call_set_name: Option<String>,
}

A call represents the determination of genotype with respect to a particular variant. It may include associated information such as quality and phasing. For example, a call might assign a probability of 0.32 to the occurrence of a SNP named rs1234 in a call set with the name NA12345.

This type is not used in any activity, and only used as part of another schema.

Fields
info: Option<HashMap<String, Vec<String>>>

A map of additional variant call information. This must be of the form map (string key mapping to a list of string values).

genotype: Option<Vec<i32>>

The genotype of this variant call. Each value represents either the value of the referenceBases field or a 1-based index into alternateBases. If a variant had a referenceBases value of T and an alternateBases value of ["A", "C"], and the genotype was [2, 1], that would mean the call represented the heterozygous value CA for this variant. If the genotype was instead [0, 1], the represented value would be TA. Ordering of the genotype values is important if the phaseset is present. If a genotype is not called (that is, a . is present in the GT string) -1 is returned.

call_set_id: Option<String>

The ID of the call set this variant call belongs to.

phaseset: Option<String>

If this field is present, this variant call's genotype ordering implies the phase of the bases and is consistent with any other variant calls in the same reference sequence which have the same phaseset value. When importing data from VCF, if the genotype data was phased but no phase set was specified this field will be set to *.

genotype_likelihood: Option<Vec<f64>>

The genotype likelihoods for this variant call. Each array entry represents how likely a specific genotype is for this call. The value ordering is defined by the GL tag in the VCF spec. If Phred-scaled genotype likelihood scores (PL) are available and log10(P) genotype likelihood scores (GL) are not, PL scores are converted to GL scores. If both are available, PL scores are stored in info.

call_set_name: Option<String>

The name of the call set this variant call belongs to.