Struct google_genomics1::VariantSet

pub struct VariantSet {
    pub reference_bounds: Option<Vec<ReferenceBound>>,
    pub reference_set_id: Option<String>,
    pub id: Option<String>,
    pub dataset_id: Option<String>,
    pub metadata: Option<Vec<VariantSetMetadata>>,
}

A variant set is a collection of call sets and variants. It contains summary statistics of those contents. A variant set belongs to a dataset. For more genomics resource definitions, see Fundamentals of Google Genomics

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch variantsets (request|response)

create variantsets (request|response)

get variantsets (response)

Fields

reference_bounds

A list of all references used by the variants in a variant set with associated coordinate upper bounds for each one.

reference_set_id

The reference set to which the variant set is mapped. The reference set describes the alignment provenance of the variant set, while the referenceBounds describe the shape of the actual variant data. The reference set's reference names are a superset of those found in the referenceBounds. For example, given a variant set that is mapped to the GRCh38 reference set and contains a single variant on reference 'X', referenceBounds would contain only an entry for 'X', while the associated reference set enumerates all possible references: '1', '2', 'X', 'Y', 'MT', etc.

id

The server-generated variant set ID, unique across all variant sets.

dataset_id

The dataset to which this variant set belongs.

metadata

The metadata associated with this variant set.